feat: m단위 계산 통일

+++ src/main/java/com/running/you_run/user/payload/response/UserGradeInfoResponse.java
@@ -5,7 +5,7 @@ import com.running.you_run.user.entity.User;
 public record UserGradeInfoResponse(
         int level,
         String grade,
-        Double totalDistance
+        long totalDistance
 ) {
     public static UserGradeInfoResponse from(User user){
         return new UserGradeInfoResponse(
